logger is working

+++ FlowerMonitor/Domain.h
@@ -33,7 +33,7 @@ typedef struct {
 	uint8_t hh;
 	uint8_t mm;
 	uint8_t ss;
-	uint8_t ml;
+	uint16_t ml;
 } Time;
 
 enum {
